Move API methods from initializer to ApplicationRecord.
